+++ src/conductor/executor/__init__.py
@@ -8,12 +8,15 @@ from conductor.executor.agent import AgentExecutor, resolve_agent_tools
 from conductor.executor.output import parse_json_output, validate_output
 from conductor.executor.script import ScriptExecutor, ScriptOutput
 from conductor.executor.template import TemplateRenderer
+from conductor.executor.wait import WaitExecutor, WaitOutput
 
 __all__ = [
     "AgentExecutor",
     "ScriptExecutor",
     "ScriptOutput",
     "TemplateRenderer",
+    "WaitExecutor",
+    "WaitOutput",
     "parse_json_output",
     "resolve_agent_tools",
     "validate_output",