+++ src/conductor/web/frontend/src/components/detail/DetailPanel.tsx
@@ -9,6 +9,7 @@ import { GateDetail } from './GateDetail';
 import { GroupDetail } from './GroupDetail';
 import { DialogEngagementPrompt } from './DialogEngagementPrompt';
 import { SubworkflowDetail } from './SubworkflowDetail';
+import { WaitDetail } from './WaitDetail';
 import { cn } from '@/lib/utils';
 
 export function DetailPanel() {
@@ -48,6 +49,8 @@ export function DetailPanel() {
     switch (node.type) {
       case 'script':
         return ScriptDetail;
+      case 'wait':
+        return WaitDetail;
       case 'set':
         return SetDetail;
       case 'human_gate':