+++ src/conductor/web/frontend/src/components/graph/WorkflowGraph.tsx
@@ -25,6 +25,7 @@ import { SetNode } from './SetNode';
 import { GateNode } from './GateNode';
 import { GroupNode } from './GroupNode';
 import { WorkflowNode } from './WorkflowNode';
+import { WaitNode } from './WaitNode';
 import { EndNode } from './EndNode';
 import { StartNode } from './StartNode';
 import { IngressNode } from './IngressNode';
@@ -42,6 +43,7 @@ const nodeTypes: NodeTypes = {
   gateNode: GateNode,
   groupNode: GroupNode,
   workflowNode: WorkflowNode,
+  waitNode: WaitNode,
   endNode: EndNode,
   startNode: StartNode,
   ingressNode: IngressNode,